fix bad json responses
this fixes some issues I found while prepping the demo such as:
- occasional text responses instead of json from some endpoint
- missing auth checks
- null response instead of an empty array
- checks for supported pg version
