+++ internal/metastore/metastore_test.go
@@ -27,39 +27,39 @@ func TestServices(t *testing.T) {
 
 	clusters := []ClusterInfo{
 		{
-			Host: "localhost",
-			Name: "db1",
-			ClusterID: generateID("db1"),
-			Username: "user1",
-			Password: "password1",
-			Port: 9001,
+			Host:       "localhost",
+			Name:       "db1",
+			ClusterID:  generateID("db1"),
+			Username:   "user1",
+			Password:   "password1",
+			Port:       9001,
 			MajVersion: 13,
 			MinVersion: 0,
-		},{
-			Host: "localhost",
-			Name: "db2",
-			ClusterID: generateID("db2"),
-			Username: "user2",
-			Password: "password2",
-			Port: 9001,
+		}, {
+			Host:       "localhost",
+			Name:       "db2",
+			ClusterID:  generateID("db2"),
+			Username:   "user2",
+			Password:   "password2",
+			Port:       9001,
 			MajVersion: 13,
 			MinVersion: 0,
-		},{
-			Host: "localhost",
-			Name: "db3",
-			ClusterID: generateID("db3"),
-			Username: "user3",
-			Password: "password3",
-			Port: 9002,
+		}, {
+			Host:       "localhost",
+			Name:       "db3",
+			ClusterID:  generateID("db3"),
+			Username:   "user3",
+			Password:   "password3",
+			Port:       9002,
 			MajVersion: 13,
 			MinVersion: 0,
-		},{
-			Host: "localhost",
-			Name: "db4",
-			ClusterID: generateID("db4"),
-			Username: "user4",
-			Password: "password4",
-			Port: 9003,
+		}, {
+			Host:       "localhost",
+			Name:       "db4",
+			ClusterID:  generateID("db4"),
+			Username:   "user4",
+			Password:   "password4",
+			Port:       9003,
 			MajVersion: 13,
 			MinVersion: 0,
 		},
