+++ testutils/testutil.go
@@ -1,0 +1,8 @@
+package testutils
+
+import "github.com/spinup-host/spinup/config"
+
+// GetConfig returns a test configuration that can be used in tests.
+func GetConfig() config.Configuration {
+	return config.Configuration{}
+}
