+++ utils/logger.go
@@ -1,4 +1,5 @@
 package utils
+
 import (
 	"fmt"
 	"log"
@@ -7,6 +8,7 @@ import (
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
+
 var Logger *zap.Logger
 
 // InitializeLogger sets up a log file to write spinup logs to.
@@ -15,7 +17,7 @@ func InitializeLogger(logDir string, fileName string) {
 
 	config := zap.NewProductionEncoderConfig()
 	if logDir == "" {
-		homeDir, _ := os.UserHomeDir();
+		homeDir, _ := os.UserHomeDir()
 		loggingFilePath = homeDir
 	} else {
 		loggingFilePath = logDir
@@ -27,7 +29,7 @@ func InitializeLogger(logDir string, fileName string) {
 	}
 
 	log.Println(fmt.Sprintf("using log file %s\n", loggingFilePath))
-    config.EncodeTime = zapcore.ISO8601TimeEncoder
+	config.EncodeTime = zapcore.ISO8601TimeEncoder
 	fileEncoder := zapcore.NewJSONEncoder(config)
 	consoleEncoder := zapcore.NewConsoleEncoder(config)
 	logFile, _ := os.OpenFile(loggingFilePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
